Don't announce via DNS-SD if KISS port not configured
When "KISSPORT 0" is used in the config file to remove the default
port, and no other port is configured, Direwolf still announces a
KISS service on port 0. This is happening because the check for a
valid KISS port was not updated when 'kiss_port' was changed from
a single value to an array (for support of multiple KISS ports),
so the test is checking against the wrong value.

Checking against the first port in the list, per this commit, is
consistent with the use of this same value for the port number that
is announced via DNS-SD. When support is added for announcing multiple
ports, this check should be updated to check for any valid port.

Fixes #386