Add new CMake option to build without DNS-SD
On startup, Direwolf states that it includes optional support for
dns-sd. However, no option is provided in the build to exclude such
support. This means that on Linux, for example, the Avahi libraries
are always required.

This change adds a CMake option to exclude DNS-SD support, for both
Linux and Mac builds. The default is to include it, retaining the
current status. To exclude it, '-DOPTIONAL_DNSSD=OFF' can be added
to the invocation of CMake.

Fixes #451