New SCHANNEL feature, like NCHANNEL but for serial TNCs
The SCHANNEL feature is essentially the same as the existing NCHANNEL
feature but allowing for the mapping of a serial TNC to a channel, as
opposed to the network TNC supported by NCHANNEL.

Configuration uses the following form in the direwolf.conf file:

    SCHANNEL <channel> <device> [<baudrate>]

The implementation is a close parallel to that for NCHANNEL, sharing
some of the code, and makes use of existing serial port handling that
hides OS differences. This results in fewer changes than might have
been expected otherwise.